requirements.txt
numpy >= 1.*
python-dotenv >= 0.*
dnspython == 2.2.1
pymongo == 4.1.1
fastapi == 0.78.0
uvicorn == 0.17.6
bcrypt == 3.2.2
pyjwt == 2.4.0
pydantic==1.9.1
certifi==2022.5.18.1
python-dateutil == 2.8.2
pyyaml == 6.0
stripe == 3.2.0
requests == 2.27.1